Fix the build

+++ beacon_chain/consensus_object_pools/blockchain_dag.nim
@@ -1247,7 +1247,7 @@ proc updateState*(
         ancestors.add(cur.bid)
 
       if cur.slot == GENESIS_SLOT or
-          (cur.slot.epoch + EPOCHS_PER_STATE_SNAPSHOT * 2 < startEpoch):
+          (cur.slot.epoch +  uint64(EPOCHS_PER_STATE_SNAPSHOT) * 2 < startEpoch):
         # We've either walked two full state snapshot lengths or hit the tail
         # and still can't find a matching state: this can happen when
         # starting the node from an arbitrary finalized checkpoint and not
